+++ ChangeLog.md
@@ -1,12 +1,14 @@
 # Revision history for patch
 
+## Unreleased
+
+* Use `commutative-semigroups` for `Commutative`, making `Additive` a
+  deprecated alias.
+
 ## 0.0.6.0 - 2022-06-10
 
 * Add `PatchOrReplacement`, patch which either is some other patch type or a
   new replacement value.
-
- * Use `commutative-semigroups` for `Commutative`, making `Additive` a
-   deprecated alias.
 
 * Support GHC 9.2