+++ ChangeLog.md
@@ -1,5 +1,10 @@
 # Revision history for patch
 
+## Unreleased
+
+* Use `commutative-semigroups` for `Commutative`, making `Additive` a
+  deprecated alias.
+
 ## 0.0.6.0 - 2022-06-10
 
 * Add `PatchOrReplacement`, patch which either is some other patch type or a